refactor: move extractor building into `scalibrplugin` package
